+++ src/main/java/com/tomsawyer/perspectives/automation/sikuli/tests/TSHtml5PreviewTestCases.java
@@ -1,3 +1,4 @@
+package com.tomsawyer.perspectives.automation.sikuli.tests;
 import java.io.File;
